pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skylapse"
version = "0.1.0"
description = "Raspberry Pi HQ Camera timelapse system with FastAPI web app"
readme = "README.md"
authors = [{ name = "Skylapse" }]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pyyaml>=6.0.1,<7.0",
  "opencv-python>=4.10,<5.0",
  "Pillow>=10.4,<11.0",
  "sqlalchemy>=2.0,<3.0",
  "paho-mqtt>=2.1,<3.0",
  "boto3>=1.34,<2.0",
]
optional-dependencies.dev = [
  "pytest>=8.3,<9.0",
  "pytest-cov>=5.0,<6.0",
  "httpx>=0.27,<0.28",
  "mypy>=1.11,<2.0",
  "ruff>=0.6,<0.7",
  "black>=24.8,<25.0",
  "types-PyYAML>=6.0.12,<7.0",
]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "C90"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
plugins = []

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = ["tests"]
